requirements.txt
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
pandas>=2.0.0
openpyxl>=3.1.0